+++ crates/plugin/service/impl/src/lib.rs
@@ -12,7 +12,6 @@ pub mod container;
 pub mod context;
 pub mod plugin_container_manager_impl;
 pub mod plugin_context_factory_impl;
-pub mod plugin_paths;
 pub mod plugin_repository_manager_impl;
 pub mod plugin_resolver_impl;
 pub mod plugin_system_impl;