+++ crates/reactive-graph/src/tooling/releases/release.rs
@@ -1,6 +1,8 @@
 use serde::Serialize;
 use tabled::Tabled;
 
+pub const TARGET_TRIPLE: &str = env!("VERGEN_CARGO_TARGET_TRIPLE");
+
 #[derive(Clone, Debug, Serialize, Tabled)]
 pub struct Release {
     #[tabled(rename = "Release Name")]
@@ -17,7 +19,7 @@ pub struct Release {
 
 impl From<&self_update::update::Release> for Release {
     fn from(release: &self_update::update::Release) -> Self {
-        let (asset_name, download_url) = match release.asset_for(env!("VERGEN_CARGO_TARGET_TRIPLE"), None) {
+        let (asset_name, download_url) = match release.asset_for(TARGET_TRIPLE, None) {
             Some(release_asset) => (release_asset.name, release_asset.download_url),
             None => ("Not available".to_string(), String::new()),
         };